Abstract countDown timer + refetch the google state once
